Add support for Test users
When a user is signed-in to Identity, and they have a valid test-user
id, they get given the 'test' mode of Stripe. Otherwise, on production,
they get the 'Live' mode as usual.

Updated Stripe config:

This uses a slightly simpler format than Membership projects, because we
don't have the DEV/UAT/PROD backend distinction here.

Public and secret keys in Stripe are rotated separately, so storing them
separately (with the public key in source control and the secret key in
S3) seems reasonable to me.