Add a filter which requires that every response has cache headers
This is borrowed from Membership, with a few tweaks to use Java 8's time
classes. Note that only HTTP 200, 301, 404 responses are cached - you
generally don't want to cache 500 errors that correspond to server crashes
for instance.


Thanks to Alex for pointing out that the documentation at:

https://www.playframework.com/documentation/2.5.x/ScalaHttpFilters

...is totally unhelpful if you're using compile-time DI. I've deleted
much of the application.conf where that same confusion occurred.