Refactor Test class to allow for canRun checks
In addition to audience segmentation, you can now create tests with or without can run criteria, which I've
called "split tests" and "conditional tests". The can run check is a
boolean function that takes a request as its only parameter. I figure
this should be enough for most use cases.